refactor: 공통 타입,상수,instance 리팩토링
