+++ src/@types/chat.ts
@@ -1,3 +1,6 @@
+import { Message } from './message';
+import { User } from './user';
+
 export type ResponseValue = Chat[];
 
 export interface Chat {
@@ -7,17 +10,4 @@ export interface Chat {
   isPrivate: boolean;
   latestMessage: Message | null;
   updatedAt: Date;
-}
-
-export interface User {
-  id: string;
-  name: string;
-  picture: string;
-}
-
-export interface Message {
-  id: string;
-  text: string;
-  userId: string;
-  createAt: Date;
 }
