--- src/constants/apiUrl.ts
@@ -1,1 +1,0 @@
-export const API_BASED_URL = process.env.REACT_APP_API_URL;
