--- src/constants/url.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = 'https://fastcampus-chat.net';
