+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './channelName';
+export * from './formatTimeToKST';
